+++ test/framework/src/main/java/org/opensearch/test/telemetry/MockTelemetry.java
@@ -10,6 +10,7 @@ package org.opensearch.test.telemetry;
 
 import org.opensearch.telemetry.Telemetry;
 import org.opensearch.telemetry.TelemetrySettings;
+import org.opensearch.telemetry.metrics.Counter;
 import org.opensearch.telemetry.metrics.MetricsTelemetry;
 import org.opensearch.telemetry.tracing.TracingTelemetry;
 import org.opensearch.test.telemetry.tracing.MockTracingTelemetry;
@@ -34,6 +35,20 @@ public class MockTelemetry implements Telemetry {
     @Override
     public MetricsTelemetry getMetricsTelemetry() {
         return new MetricsTelemetry() {
+            @Override
+            public Counter createCounter(String name, String description, String unit) {
+                return null;
+            }
+
+            @Override
+            public Counter createUpDownCounter(String name, String description, String unit) {
+                return null;
+            }
+
+            @Override
+            public void close() {
+
+            }
         };
     }
 }